Return redirect responses, render exceptions a little nicer
